apps/backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
rq==1.16.2
PyJWT==2.8.0
cryptography==42.0.5
httpx==0.27.0
